Don't always try to create the service image.
This patch makes it possible to create a service_instance resource
without also attempting to upload the image to glance.

This is required because the glance_image resource must be run on the
same host as the glance api service, meaning that it will fail when
trying to create a stand alone manila host.

Defaults to true in order to retain backwards compatibility.

Change-Id: Id863272652db2b01e8d86b94cfd417fa78f7ea1b